+++ docs/changelog.rst
@@ -22,6 +22,10 @@ Changelog
 * Fixed Console's text color in dark/light mode.
 * Fixed mouse click in the Editor View not selecting the correct sprite in some circumnstances.
 
+Lua API**
+
+* Added ``Device.screenshot()`` and ``screenshot()`` callback, see :doc:`lua_api` for details.
+
 0.49.0 --- 27 Nov 2023
 ----------------------